+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- feat(shortcuts): enhance shortcut handling and add new utility functions. (`c71daed`)
 - fix(catalog): update plugin version to v1.0.7 and enhance ScreenshotCarousel functionality. (`71ee623`)
 - feat(plugin-detail-modal): enhance screenshot handling in PluginDetailModal. (`a07c307`)